Reworked the project
The previous setup for JWT was very complex to work with, reworked to use the example code provided by Girts as it is more comprehensible. Added API to register a user. JwtRequestFilter probably needs to be tweaked, as it blocks all requests except for authorization, making it currently impossible to register a user. Therefore the method has not been tested to see if it works.